+++ hexa/catalog/tests/test_views.py
@@ -1,5 +1,3 @@
-from unittest import skip
-
 from django import test
 from django.db.models import QuerySet
 from django.http import JsonResponse
@@ -15,7 +13,7 @@ from hexa.plugins.connector_dhis2.models import (
     InstancePermission,
     OrganisationUnit,
 )
-from hexa.plugins.connector_s3.models import Bucket, BucketPermission, Object
+from hexa.plugins.connector_s3.models import Bucket, BucketPermission
 from hexa.user_management.models import Team, User
 
 
@@ -284,24 +282,6 @@ class CatalogTest(test.TestCase):
         response = self.client.get(f"{reverse('catalog:quick_search')}?query=anc")
         self.assertEqual(response.status_code, 200)
         self.assertEqual(0, len(response.json()["results"]))
-
-    @skip("wait for orphan to disapear")
-    def test_catalog_search_orphan(self):
-        """The search should not return orphan objects"""
-        object = Object.objects.create(
-            bucket=self.BUCKET,
-            key="test-orphanXXAAAXXXXAAAAXXXX",
-            parent_key="",
-            size=100,
-            storage_class="STANDARD",
-            type="file",
-            orphan=True,
-        )
-        self.client.force_login(self.USER_KRISTEN)
-        response = self.client.get(f"{reverse('catalog:quick_search')}?query=orphan")
-        self.assertEqual(response.status_code, 200)
-        for result in response.json()["results"]:
-            self.assertTrue(result["external_name"] != "test-orphanXXAAAXXXXAAAAXXXX")
 
     def test_catalog_search_should_validate_filter_input(self):
         """Typing an invalid/corrupted filter query should not result in an error 500"""